Processing Refactor (#409)
* refactor: have proversync cache proofs upon tree update

* refactor: start introducing acceptableRoot into replica

* refactor: set confirmAt to 1 for initial root

* refactor: remove previous from replica trait

* refactor: processor no longer bails on missing proof

* tests: fix the tests

* refactor: remove last use of tokio::time::interval

* refactor: remove redundant condition in replica

+++ rust/optics-core/src/accumulator/incremental.rs
@@ -75,7 +75,7 @@ impl IncrementalMerkle {
 
     /// Verify a incremental merkle proof of inclusion
     pub fn verify(&self, proof: &Proof) -> bool {
-        let computed = IncrementalMerkle::branch_root(proof.leaf, proof.path, proof.index);
+        let computed = IncrementalMerkle::branch_root(proof.leaf, proof.path, proof.index as usize);
         computed == self.root()
     }
 }
